requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
psycopg2==2.9.9
WTForms==3.1.1